services/monitoring_service/requirements.txt
# Monitoring Service dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
httpx>=0.24.0

# Service discovery
nacos-sdk-python>=0.1.14

# Configuration
python-dotenv>=1.0.0

# Langfuse observability
langfuse>=2.0.0

# Prometheus metrics export
prometheus-client>=0.19.0
